crud proyectos terminado + chequeo errores

+++ services/projects_service.py
@@ -8,14 +8,21 @@ def add_project(project):
 def get_all_projects():
     return Project.query.all()
 
+def get_project(id):
+    return Project.query.get(id)
+
 def delete_project(id):
     project = Project.query.get(id)
-    db.session.delete(project)
-    db.session.commit()
+    if project:
+        db.session.delete(project)
+        db.session.commit()
+        return True
+    False
 
 def update_project(id, data):
     project = Project.query.get(id)
-    project.name = data['name']
-    project.cost = data['cost']
-    project.created_at = data['created_at']
-    db.session.commit()
+    if project:
+        project.update_data(data)
+        db.session.commit()
+        return project
+    return False
